# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assessment System API",
    "description": "A comprehensive API for managing assessments, submissions, and reviews in a sustainability assessment system.\n\n## Authentication\nAll endpoints require JWT authentication. Include the token in the Authorization header:\n`Authorization: Bearer <your-jwt-token>`\n\n## Versioning\nThe API supports response versioning to track changes and resolve conflicts.\n",
    "contact": {
      "name": "Assessment System Team",
      "email": "support@assessment-system.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/category-catalog": {
      "get": {
        "tags": ["Category Catalog"],
        "summary": "List category catalogs",
        "description": "Get all active category catalogs",
        "operationId": "get_category_catalogs",
        "responses": {
          "200": {
            "description": "List of active category catalogs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCatalogListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Category Catalog"],
        "summary": "Create category catalog",
        "description": "Create a new category catalog entry (Admin only)",
        "operationId": "create_category_catalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Category catalog created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCatalogResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/category-catalog/{category_catalog_id}": {
      "get": {
        "tags": ["Category Catalog"],
        "summary": "Get category catalog",
        "description": "Get a specific category catalog by ID",
        "operationId": "get_category_catalog",
        "parameters": [
          {
            "name": "category_catalog_id",
            "in": "path",
            "description": "Category Catalog ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category catalog found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCatalogResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Category Catalog"],
        "summary": "Update category catalog",
        "description": "Update a category catalog entry (Admin only)",
        "operationId": "update_category_catalog",
        "parameters": [
          {
            "name": "category_catalog_id",
            "in": "path",
            "description": "Category Catalog ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Category catalog updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCatalogResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Category Catalog"],
        "summary": "Delete category catalog",
        "description": "Delete a category catalog entry by id (Admin only)",
        "operationId": "delete_category_catalog",
        "parameters": [
          {
            "name": "category_catalog_id",
            "in": "path",
            "description": "Category Catalog ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category catalog deleted successfully"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/organizations/{keycloak_organization_id}/categories": {
      "get": {
        "tags": ["Organizations"],
        "summary": "Get organization categories",
        "description": "Get all categories assigned to a specific organization",
        "operationId": "get_organization_categories",
        "parameters": [
          {
            "name": "keycloak_organization_id",
            "in": "path",
            "description": "Keycloak Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of organization categories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationCategoryListResponse"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/organizations/{keycloak_organization_id}/categories/assign": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Assign categories to organization",
        "description": "Assign a list of categories to an organization. This will replace existing assignments.",
        "operationId": "assign_categories_to_organization",
        "parameters": [
          {
            "name": "keycloak_organization_id",
            "in": "path",
            "description": "Keycloak Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCategoriesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categories assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationCategoryListResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/organizations/{keycloak_organization_id}/categories/{organization_category_id}": {
      "put": {
        "tags": ["Organizations"],
        "summary": "Update organization category weight",
        "description": "Update the weight of a single category for an organization.",
        "operationId": "update_organization_category",
        "parameters": [
          {
            "name": "keycloak_organization_id",
            "in": "path",
            "description": "Keycloak Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organization_category_id",
            "in": "path",
            "description": "Organization Category ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationCategory"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CategoryCatalog": {
        "required": ["category_catalog_id", "name", "template_id", "is_active", "created_at", "updated_at"],
        "type": "object",
        "properties": {
          "category_catalog_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "template_id": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCategoryCatalogRequest": {
        "required": ["name", "template_id"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "template_id": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "UpdateCategoryCatalogRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "CategoryCatalogResponse": {
        "required": ["category_catalog"],
        "type": "object",
        "properties": {
          "category_catalog": {
            "$ref": "#/components/schemas/CategoryCatalog"
          }
        }
      },
      "CategoryCatalogListResponse": {
        "required": ["category_catalogs"],
        "type": "object",
        "properties": {
          "category_catalogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryCatalog"
            }
          }
        }
      },
...
      "AssignCategoriesRequest": {
        "type": "object",
        "required": ["category_catalog_ids"],
        "properties": {
          "category_catalog_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "List of Category Catalog IDs to assign to the organization."
          }
        }
      },
      "UpdateOrganizationCategoryRequest": {
        "type": "object",
        "required": ["weight"],
        "properties": {
          "weight": {
            "type": "number",
            "format": "float",
            "description": "The weight for this category for this organization (0-100)."
          }
        }
      },
      "OrganizationCategory": {
        "type": "object",
        "required": [
          "organization_category_id",
          "organization_id",
          "category_catalog_id",
          "weight",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "organization_category_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string"
          },
          "category_catalog_id": {
            "type": "string",
            "format": "uuid"
          },
          "weight": {
            "type": "number",
            "format": "float"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrganizationCategoryListResponse": {
        "type": "object",
        "required": ["categories"],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationCategory"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}